fix: remove empty lines on top of man page
The added reuse license and copyright lines as comments did create
empty lines on top. Heading the comment with a dot (.) That character
at the beginning of a line introduces a request or macro. Followed by a
comment nothing will happen but heading empty lines are gone. One is
still there.

 - .\" instead \"

Signed-off-by: Conrad Zelck <git@simpel.cc>